api-server/requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.7
Mako==1.3.3
MarkupSafe==2.1.5
playwright==1.43.0
psycopg2==2.9.9
pydantic==2.7.0
pydantic_core==2.18.1
pyee==11.1.0
python-dotenv==1.0.1
PyYAML==6.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
psycopg2-binary~=2.9.9